+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 * Magical Glass can no longer be placed
 * (API) Removed deprecated "SlimefunBlockHandler"
 * Removed Automated Crafting Chamber
+* Memory and performance improvements for Cargo and Energy networks
 
 #### Fixes
 * Fixed #2987